+++ backend/src/__tests__/products.controller.test.js
@@ -1,6 +1,6 @@
-const ProductsController = require('../Controllers/productController');
+const ProductsController = require('../controllers/product.controller');
 const mysql = require('mysql2');
-const config = require('../Config/appConfig');
+const config = require('../config/app.config');
 
 // Mock MySQL connection
 jest.mock('mysql2');
