+++ backend/src/__tests__/wishlist.controller.test.js
@@ -1,6 +1,6 @@
-const WishlistController = require('../Controllers/wishlistController');
+const WishlistController = require('../controllers/wishlist.controller');
 const mysql = require('mysql2');
-const config = require('../Config/appConfig');
+const config = require('../config/app.config');
 
 // Mock MySQL connection
 jest.mock('mysql2');
